Support watching independent files
As explained in the docs, JDK's watch service cannot watch independent
files and instead can only watch directories. To simulate watching
independent files, we watch their parent directories in a non-recursive
manner. This PR implements the bits to simulate this behaviour.

We only register non-recursive paths if they have not been registered
before or if they are not contained in a path that is already been
watched.